requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data
h5py>=3.9.0
PyYAML>=6.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Dashboard
streamlit>=1.28.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# MLflow (experiment tracking)
mlflow>=2.8.0

# ONNX (model optimization)
onnx>=1.14.0
onnxruntime>=1.16.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0